adding script
Using a local copy of the semgrep-rules repository, this script recurses through all .yaml files creating a list of the paths to each rule file. Then three files are generated which search each path in the rules file for three specified categories, "message:", "owasp:", and "languages". The paths to rules containing a minimum one of those matches are added to a file. Then those three categories are searched for the keyword entered by the user.